+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 ## [8.0.0-alpha-003] - 2026-03-03
 
+### Changed
+
+- Improve formatting of query expressions with join/on/into clauses. [#3156](https://github.com/fsprojects/fantomas/issues/3156)
+
 ### Fixed
 
 - End of line comments after infix operators are preserved correctly. [#2287](https://github.com/fsprojects/fantomas/issues/2287)